# docs/swagger.yaml
basePath: /
definitions:
  rest_service_internal_subscriptionService.RequestBody:
    properties:
      end_date:
        description: тоже "MM-YYYY"
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        description: формат "MM-YYYY"
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  rest_service_internal_subscriptionService.Subscription:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: support@example.com
    name: Поддержка API
  description: REST API для управления подписками.
  termsOfService: http://example.com/terms/
  title: Subscription API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Возвращает список всех подписок
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/rest_service_internal_subscriptionService.Subscription'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создает подписку с переданными параметрами
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/rest_service_internal_subscriptionService.RequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest_service_internal_subscriptionService.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать новую подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Удаляет подписку с указанным ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить подписку по ID
      tags:
      - subscriptions
    get:
      description: Возвращает подписку по уникальному идентификатору
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest_service_internal_subscriptionService.Subscription'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить подписку по ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет данные подписки с указанным ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: string
      - description: Обновленные данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/rest_service_internal_subscriptionService.RequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest_service_internal_subscriptionService.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить подписку по ID
      tags:
      - subscriptions
  /subscriptions/amountSubscriptions:
    get:
      description: Возвращает общую сумму подписок за указанный период с учетом фильтров
      parameters:
      - description: Дата начала (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Дата окончания (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: ID пользователя
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        in: query
        name: name_service
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить сумму подписок по фильтрам
      tags:
      - subscriptions
swagger: "2.0"